Texture.h: Limiting materials vector to use unsigned 8-bit integers instead of 64 bit integers for sake of memory economy. Who needs more than 254 materials anyways. - Internally, mat_none / '-' is stored as 255 (uint8_t), but externally (int) as -1. - Now we can encode and decode the complete set of available materials as a string which makes it much safer, reliable and "feature-complete". - Fixing (not yet encountered) range bugs in subset(). - Because of how materials are propagated throughout the Termin8or API, we unfortunately have to change the type of material variables everywhere from int to uint8_t. It is what it is.

+++ include/Termin8or/physics/dynamics/RigidBody.h
@@ -51,8 +51,8 @@ namespace t8x
     AABB<int> curr_sprite_aabb;
     AABB<float> curr_aabb;
     bool_vector curr_inertia_mask, curr_coll_mask;
-    std::vector<int> inertia_materials { 1 };
-    std::vector<int> collision_materials { 1 };
+    std::vector<uint8_t> inertia_materials { 1 };
+    std::vector<uint8_t> collision_materials { 1 };
     std::vector<Vec2> surface_normals;
     
     bool enable_sleeping = false;
@@ -154,8 +154,8 @@ namespace t8x
       float e = 0.8f, float dyn_friction = 0.f,
       std::optional<float> crit_speed_r = std::nullopt,
       std::optional<float> crit_speed_c = std::nullopt,
-      const std::vector<int>& inertia_mats = { 1 },
-      const std::vector<int>& coll_mats = { 1 })
+      const std::vector<uint8_t>& inertia_mats = { 1 },
+      const std::vector<uint8_t>& coll_mats = { 1 })
       : sprite(s)
       , mass(rb_mass)
       , inv_mass(rb_mass == 0.f ? 0.f : 1.f / rb_mass)
